deactivate print tracking sentence
